+++ components/landing.tsx
@@ -4,102 +4,101 @@ import '../app/css/additional-styles/landing.css';
 import ImageOne from '../public/images/demo.jpg';
 import Image from "next/image";
 
-export default function Landing()
-{
-    return(
-        <>
+export default function Landing() {
+  return (
+    <>
 
       <div className="container place-items-center font-bold pt-20 pb-10">
         <h2 className="text-5xl text-white-800 text-center">
-           Activities
+          Activities
         </h2>
         <h5 className="text-2xl text-white-800 text-center">
-           We organise lots of student centric activities
+          We organise lots of student centric activities
         </h5>
       </div>
 
 
-      
-    <div className="flex flex-col md:flex-row w-full h- ">
-      <div className="highlight flex-1 flex items-center justify-center alip-2 bg-black-800 text-green-500 ">
-         <div className="p-8 md:p-16 max-w-lg text-center ">
-             <h1 className="text-3xl font-bold mb-4 ">
-              COMPETITIVE PROGRAMMING
-             </h1>
-             <h2 className="text-2xl  font-bold mb-4">
-              PB Hustle
-             </h2>
-             <p className="text-lg text-center md:text-center">
-             Point Blank has organized over 40+ editions of its PB Hustle coding contest,
-              where participants solve 5-7 increasingly difficult problems in their preferred 
-              programming language. The contest aims to enhance the programming culture in colleges
-               and help teams qualify for the ACM ICPC. Impressively, DSCE's leading programmers have
-                risen through this platform, with participation in the CodeChef Long Challenge expanding 
-                from just 3 to over 70+ participants.
-             </p>
-           </div>
-         </div>
-        <div className=" highlight flex-1 w-full flex items-center justify-center bg-black-900">
-        <Image src={ImageOne} alt='' className=''/>
-        </div>
-    </div>
 
-    <div className="flex flex-col md:flex-row w-full h-">
+      <div className="flex flex-col md:flex-row w-full h- ">
+        <div className="highlight flex-1 flex items-center justify-center alip-2 bg-black-800 text-green-500 ">
+          <div className="p-8 md:p-16 max-w-lg text-center ">
+            <h1 className="text-3xl font-bold mb-4 ">
+              COMPETITIVE PROGRAMMING
+            </h1>
+            <h2 className="text-2xl  font-bold mb-4">
+              PB Hustle
+            </h2>
+            <p className="text-lg text-center md:text-center">
+              Point Blank has organized over 40+ editions of its PB Hustle coding contest,
+              where participants solve 5-7 increasingly difficult problems in their preferred
+              programming language. The contest aims to enhance the programming culture in colleges
+              and help teams qualify for the ACM ICPC. Impressively, DSCE's leading programmers have
+              risen through this platform, with participation in the CodeChef Long Challenge expanding
+              from just 3 to over 70+ participants.
+            </p>
+          </div>
+        </div>
+        <div className=" highlight flex-1 w-full flex items-center justify-center bg-black-900">
+          <Image src={ImageOne} alt='' className='' />
+        </div>
+      </div>
+
+      <div className="flex flex-col md:flex-row w-full h-">
         <div className="highlight flex-1 flex items-center justify-center bg-black-500 h-full">
-          <Image src={ImageOne} alt='' className=''/>
+          <Image src={ImageOne} alt='' className='' />
         </div>
         <div className=" highlight flex-1 flex items-center justify-center alip-8 bg-black-800 text-green-500">
           <div className=" p-8 md:p-16 max-w-lg text-center ">
-             <h1 className="text-3xl  font-bold mb-4 ">
+            <h1 className="text-3xl  font-bold mb-4 ">
               DEVELOPMENT
-             </h1>
-             <h2 className="text-2xl font-bold mb-4">
+            </h1>
+            <h2 className="text-2xl font-bold mb-4">
               PB Chronicals
-             </h2>
-             <p className="text-lg text-center md:text-center ">
-             Point Blank has hosted over 100+ workshops, covering a wide range 
-             of topics including open source, DevOps, machine learning, placement 
-             preparation, data structures and algorithms, and mobile and web development, 
-             among others. These workshops are typically conducted in an informal and unscripted 
-             manner, though we do document some of our most significant sessions. 
-             One notable example is our primer on F/OSS development.
-             </p>
-           </div>
-         </div>
+            </h2>
+            <p className="text-lg text-center md:text-center ">
+              Point Blank has hosted over 100+ workshops, covering a wide range
+              of topics including open source, DevOps, machine learning, placement
+              preparation, data structures and algorithms, and mobile and web development,
+              among others. These workshops are typically conducted in an informal and unscripted
+              manner, though we do document some of our most significant sessions.
+              One notable example is our primer on F/OSS development.
+            </p>
+          </div>
         </div>
+      </div>
 
-    <div className="flex flex-col md:flex-row w-full h-">
-      <div className="highlight flex-1 flex items-center justify-center alip-8 bg-black-800 text-green-500">
-         <div className=" p-8 md:p-16 max-w-lg text-center ">
-             <h1 className="text-3xl  font-bold mb-4 ">
-             HACKATHONS
-             </h1>
-             <h2 className="text-2xl  font-bold mb-4">
+      <div className="flex flex-col md:flex-row w-full h-">
+        <div className="highlight flex-1 flex items-center justify-center alip-8 bg-black-800 text-green-500">
+          <div className=" p-8 md:p-16 max-w-lg text-center ">
+            <h1 className="text-3xl  font-bold mb-4 ">
+              HACKATHONS
+            </h1>
+            <h2 className="text-2xl  font-bold mb-4">
               Smart India Hackathon
-             </h2>
-              <p className="text-lg text-center md:text-center">
-              Each year, we organize the internal 
-              round of the Smart India Hackathon. In the 2020 edition, over 300+ 
-              individuals from DSCE participated. Two of our teams advanced to the 
+            </h2>
+            <p className="text-lg text-center md:text-center">
+              Each year, we organize the internal
+              round of the Smart India Hackathon. In the 2020 edition, over 300+
+              individuals from DSCE participated. Two of our teams advanced to the
               finals, with one emerging as the winner of the software edition.
               Along with this, teams from Point Blank have won hackathons all across
-                the city and country.
-             </p>
-           </div>
-         </div>
-        <div className=" highlight flex-1 flex items-center justify-center bg-black-900">
-        <Image src={ImageOne} alt='' className=''/>
+              the city and country.
+            </p>
+          </div>
         </div>
-    </div>
+        <div className=" highlight flex-1 flex items-center justify-center bg-black-900">
+          <Image src={ImageOne} alt='' className='' />
+        </div>
+      </div>
 
-    <div className="flex flex-col md:flex-row w-full h-">
-      
+      <div className="flex flex-col md:flex-row w-full h-">
+
         <div className="highlight flex-1 flex items-center justify-center bg-black-900">
-        <Image src={ImageOne} alt='' className=''/>
+          <Image src={ImageOne} alt='' className='' />
         </div>
 
         <div className="highlight flex-1 flex items-center justify-center alip-8 bg-black-800 text-green-500">
-         <div className=" p-8 md:p-16 max-w-lg text-center ">
+          <div className=" p-8 md:p-16 max-w-lg text-center ">
             <h1 className="text-3xl font-bold mb-4 ">
               CYBER SECURITY
             </h1>
@@ -107,26 +106,26 @@ export default function Landing()
               PB CTF
             </h2>
             <p className="text-lg text-center md:text-center">
-            We organize workshops and sessions on various topics in cybersecurity, 
-            including hands-on practice on different platforms. In 2023, we launched 
-            the first in-house Capture The Flag event, PBCTF, which attracted over 70+ participants.
-             </p>
-           </div>
-         </div>
-      </div>                                               
+              We organize workshops and sessions on various topics in cybersecurity,
+              including hands-on practice on different platforms. In 2023, we launched
+              the first in-house Capture The Flag event, PBCTF, which attracted over 70+ participants.
+            </p>
+          </div>
+        </div>
+      </div>
 
 
       <div className="container place-items-center font-bold pt-20 pb-10">
-         <h2 className="text-5xl text-white-800 text-center">
-           Events
-         </h2>
-         <h5 className="text-2xl text-white-800 text-center">
-           We organise lots of student centric activities
-         </h5>
+        <h2 className="text-5xl text-white-800 text-center">
+          Events
+        </h2>
+        <h5 className="text-2xl text-white-800 text-center">
+          We organise lots of student centric activities
+        </h5>
       </div>
 
       <div className="view">
-       <EventComponent></EventComponent>
+        <EventComponent></EventComponent>
       </div>
 
 
@@ -137,18 +136,18 @@ export default function Landing()
       </div>
 
       <div className="view">
-          
+
         <CardComponents></CardComponents>
-      
+
       </div>
 
       <div className="view">
-        
+
         <CardComponents></CardComponents>
 
       </div>
-  </>
-    )
+    </>
+  )
 }
 
 
@@ -160,10 +159,10 @@ const leads = [
 ];
 
 const eventCard = [
-  { id:4 ,textt:"Back of Card", textb:"Additional info on the back of the card"},
-  { id:5 ,textt:"Back of Card", textb:"Additional info on the back of the card"},
-  { id:6 ,textt:"Back of Card", textb:"Additional info on the back of the card"},
-  
+  { id: 4, textt: "Back of Card", textb: "Additional info on the back of the card" },
+  { id: 5, textt: "Back of Card", textb: "Additional info on the back of the card" },
+  { id: 6, textt: "Back of Card", textb: "Additional info on the back of the card" },
+
 ];
 
 const CardComponents = () => {
@@ -188,20 +187,20 @@ const CardComponents = () => {
 const EventComponent = () => {
   return (
     <>
-    {eventCard.map((ec) => (
-      <div>
-         <div  key={ec.id} 
-         className="event-card">
-         <div className="card">
-           <div className="front"></div>
+      {eventCard.map((ec) => (
+        <div>
+          <div key={ec.id}
+            className="event-card">
+            <div className="card">
+              <div className="front"></div>
               <div className="back">
-                 <h2>{ec.textt}</h2>
-                 <p className='text-center'>{ec.textb}</p>
+                <h2>{ec.textt}</h2>
+                <p className='text-center'>{ec.textb}</p>
               </div>
-           </div>
+            </div>
+          </div>
         </div>
-      </div>
-    ))}
+      ))}
     </>
   )
 }
